+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Fixed using the onChange prop - always use the current function thanks to useRef
+
 ## [1.0.1] - 2024-04-30
 
 - Fixed double rendering of applet - added clean-up function for rendering hook